Created logger machine. Transitions between `Started` and `Stopped`, and performs all tasks required for those transitions (start new file, reset analog value buffer, update display...). A timer fires an event every `cardInterval` mS, but it looks like internal events can only trigger a state change, not perform an action. Actions are only performed upon changing state.
Also created a `RunningAverage` class to store current analog value
buffer. When the logger machine `timer_sample` triggers, it will grab
`_analogValue.average()`, write it to the card, and reset it.

TODO:
Should actually offload updating of the display to an onChange method in
next commit.
Need to make a new state for writing sample, so timer can trigger it, it
can fire change/log events.
Need to make machine for Wifi connection.
Need to add connections between logger and Wifi to menu.